Struct Refactor and Major Version Bump
There were a few things about how I wrote this yesterday that kind of
bugged me when I started playing around with the library. The biggest
annoyance was having the _fourby_ on the client instead of the query --
it just seemed out of place there.

There were, upon review, tons of documentation errors and the example
code wasn't even close to correct. I updated the readme and docstrings
and changed the tests to doc tests. Amusingly, this upped the test
coverage a little.

This is now in a good place, so I'm cutting it as a 1.0 release.